+++ packages/ketcher-macromolecules/src/components/preview/components/AmbiguousMonomerPreview/AmbiguousMonomerPreview.tsx
@@ -1,4 +1,3 @@
-import { AmbiguousMonomerPreviewState } from '../../../../state/common';
 import { useMemo } from 'react';
 import styled from '@emotion/styled';
 import {
@@ -8,6 +7,10 @@ import {
   Header,
   RatioBar,
 } from './AmbiguousMonomerPreview.styles';
+// Usage of aliases instead of relative import path causes error in the following import
+// because this component used directly from ketcher-react package and there are no such aliases in that package.
+// It needs to create shared components library and use it in both packages.
+import { AmbiguousMonomerPreviewState } from '../../../../state/types';
 
 interface Props {
   className?: string;
