+++ packages/ketcher-macromolecules/src/components/preview/components/PresetPreview/PresetPreview.tsx
@@ -24,11 +24,12 @@ import {
 } from './PresetPreview.styles';
 import { preview } from '../../../../constants';
 import styled from '@emotion/styled';
-import { PresetPreviewState, selectShowPreview } from 'state/common';
+import { selectShowPreview } from 'state/common';
 import { IconName } from 'ketcher-react';
 import useIDTAliasesTextForPreset from '../../hooks/useIDTAliasesTextForPreset';
 import { useAppSelector } from 'hooks';
 import IDTAliases from '../IDTAliases/IDTAliases';
+import { PresetPreviewState } from 'state';
 
 const icons: Extract<IconName, 'sugar' | 'base' | 'phosphate'>[] = [
   'sugar',
