+++ packages/ketcher-macromolecules/src/state/common/editorSlice.ts
@@ -15,61 +15,11 @@
  ***************************************************************************/
 
 import { createSlice, PayloadAction, Slice } from '@reduxjs/toolkit';
-import {
-  AmbiguousMonomerType,
-  AttachmentPointsToBonds,
-  CoreEditor,
-  IKetIdtAliases,
-  MonomerItemType,
-  PolymerBond,
-} from 'ketcher-core';
-import { RootState } from 'state';
-import { PreviewStyle, PreviewType } from 'state/types';
+import { CoreEditor } from 'ketcher-core';
+import { EditorStatePreview, RootState } from 'state';
+import { PreviewType } from 'state/types';
 import { ThemeType } from 'theming/defaultTheme';
 import { DeepPartial } from '../../types';
-
-interface BasePreviewState {
-  readonly type: PreviewType;
-  readonly style?: PreviewStyle;
-}
-
-export interface MonomerPreviewState extends BasePreviewState {
-  readonly type: PreviewType.Monomer;
-  readonly monomer: MonomerItemType | undefined;
-  readonly attachmentPointsToBonds?: AttachmentPointsToBonds;
-}
-
-export enum PresetPosition {
-  Library = 'library',
-  ChainStart = 'chainStart',
-  ChainMiddle = 'chainMiddle',
-  ChainEnd = 'chainEnd',
-}
-
-export interface PresetPreviewState extends BasePreviewState {
-  readonly type: PreviewType.Preset;
-  readonly monomers: ReadonlyArray<MonomerItemType | undefined>;
-  readonly position: PresetPosition;
-  readonly name?: string;
-  readonly idtAliases?: IKetIdtAliases;
-}
-
-export interface BondPreviewState extends BasePreviewState {
-  readonly type: PreviewType.Bond;
-  readonly polymerBond: PolymerBond;
-}
-
-export interface AmbiguousMonomerPreviewState extends BasePreviewState {
-  readonly type: PreviewType.AmbiguousMonomer;
-  readonly monomer: AmbiguousMonomerType;
-  readonly presetMonomers?: ReadonlyArray<MonomerItemType | undefined>;
-}
-
-export type EditorStatePreview =
-  | MonomerPreviewState
-  | PresetPreviewState
-  | BondPreviewState
-  | AmbiguousMonomerPreviewState;
 
 // TODO: Looks like we do not use `isReady`. Delete?
 interface EditorState {
